Polish settings and data UI
