+++ client/src/components/ToggleSwitch.jsx
@@ -22,7 +22,7 @@ export default function ToggleSwitch({ enabled, onChange, disabled, ariaLabel, s
   );
   const trackClass = `relative inline-flex ${s.track} items-center rounded-full transition-colors shrink-0 ${
     enabled ? activeColor : 'bg-gray-600'
-  } ${disabled ? 'opacity-50' : ''} ${className}`;
+  } border border-port-border/60 shadow-sm ${disabled ? 'opacity-50' : ''} ${className}`;
 
   // Decorative mode renders as a <span> so it can sit inside another <button>
   // without producing invalid nested-button HTML (which breaks tap on iOS Safari).
