+++ client/src/pages/Instances.jsx
@@ -15,21 +15,14 @@ import socket from '../services/socket';
 import {
   getInstances, updateSelfInstance, addPeer, updatePeer,
   removePeer, connectPeer, probePeer, getTailnetInfo, provisionTailnetCert,
+  getNetworkExposure,
   listPeerSubscriptions,
-  getTombstoneSweepStatus, sweepTombstonesNow,
 } from '../services/api';
 import PeerAppsList from '../components/instances/PeerAppsList';
 import PeerAgentsSection from '../components/instances/PeerAgentsSection';
 import { SchemaGapBadge } from '../components/instances/SchemaGapBadge';
 import { timeAgo } from '../utils/formatters';
 import { useLocalStorageBool } from '../hooks/useLocalStorageBool';
-import { useAsyncAction } from '../hooks/useAsyncAction';
-
-// Plural labels for tombstone record kinds — shared between the in-page
-// TombstoneGcSection toast and `scripts/gc-tombstones-now.js` (the CLI
-// echoes the same wording). The dedupe in the rendering call sites handles
-// the series/issue cohort coupling.
-const TOMBSTONE_KIND_PLURAL = { universe: 'universes', series: 'series', issue: 'issues', mediaCollection: 'media collections' };
 
 const STATUS_COLORS = {
   online: 'text-port-success',
@@ -92,7 +85,7 @@ function HealthSummary({ health, version }) {
   );
 }
 
-function TailnetHelpBanner({ tailnetInfo }) {
+function TailnetHelpBanner({ tailnetInfo, networkExposure }) {
   const [collapsed, setCollapsed] = useLocalStorageBool('portos-tailnet-help-collapsed', false);
   const [provisioning, setProvisioning] = useState(false);
   const [provisionResult, setProvisionResult] = useState(null);
@@ -118,11 +111,49 @@ function TailnetHelpBanner({ tailnetInfo }) {
 
   const ToneIcon = status.tone === 'ok' ? CheckCircle2 : AlertCircle;
   const toneClass = status.tone === 'ok' ? 'text-port-success' : 'text-port-warning';
+  const trustedHttpsHost = networkExposure?.httpsEnabled && networkExposure?.cert?.mode === 'tailscale'
+    ? networkExposure.cert.tailscaleHost || tailnetInfo?.self || null
+    : null;
+  const trustedHttpsPort = networkExposure?.bind?.port || 5555;
+  const trustedHttpsUrl = trustedHttpsHost ? `https://${trustedHttpsHost}:${trustedHttpsPort}` : null;
 
   // Only offer the one-click provision button when Tailscale is actually
   // detected and we have a MagicDNS hostname for this instance — otherwise
   // the API call will fail with the same "enable MagicDNS first" guidance.
   const canProvision = !!tailnetInfo?.self;
+
+  if (trustedHttpsHost) {
+    return (
+      <div className="bg-port-card border border-port-border rounded-xl px-4 py-3">
+        <div className="flex items-center gap-3 flex-wrap">
+          <Lock size={16} className="text-port-success shrink-0" />
+          <div className="flex-1 min-w-[220px]">
+            <div className="flex items-center gap-2 flex-wrap">
+              <span className="text-sm font-medium text-white">Tailnet DNS &amp; trusted HTTPS</span>
+              <Pill tone="success" size="xs" bordered={false} icon={CheckCircle2}>
+                Running on Tailscale HTTPS
+              </Pill>
+            </div>
+            <p className="text-xs text-gray-400 mt-1">
+              Currently running on HTTPS with Tailscale DNS.
+            </p>
+          </div>
+          {trustedHttpsUrl && (
+            <a
+              href={trustedHttpsUrl}
+              target="_blank"
+              rel="noreferrer"
+              className="inline-flex items-center gap-1.5 text-xs text-port-accent hover:text-port-accent/80 font-mono"
+              title={`Open ${trustedHttpsUrl}`}
+            >
+              {trustedHttpsHost}
+              <ArrowUpRight size={12} />
+            </a>
+          )}
+        </div>
+      </div>
+    );
+  }
 
   return (
     <div className="bg-port-card border border-port-border rounded-xl">
@@ -969,111 +1000,12 @@ function PeerCard({ peer, onRefresh, syncStatus, tailnetInfo }) {
   );
 }
 
-// Manual "GC tombstones now" trigger. The 60s orchestrator sweep keeps its
-// 24h grace — this button passes graceMs:0 so a user who just mass-deleted
-// doesn't wait a day for the bytes to leave disk. The button disables only
-// when every cohort is refused (snapshot-mode peer with no per-record sub
-// in both `universe` and `pipeline`): partial progress is still useful.
-function TombstoneGcSection() {
-  const [refused, setRefused] = useState(null); // null = still loading
-
-  // Subscribe to peer-state changes so toggling a snapshot-mode peer's
-  // sync category (or disabling it) re-evaluates the refusal status
-  // without a page reload — otherwise the button stays stuck disabled
-  // even after the user has resolved the underlying refusal condition.
-  //
-  // `instances:peers:updated` fires from `updatePeer` BEFORE the async
-  // `autoSubscribePeerToAllRecords` backfill creates per-record subs, so an
-  // immediate fetch sees the pre-backfill state. Schedule a follow-up fetch
-  // after a short delay to capture the post-backfill subs; debounce so a
-  // rapid burst of peer updates collapses to one delayed fetch.
-  useEffect(() => {
-    let cancelled = false;
-    let backfillTimer = null;
-    const fetchStatus = () => {
-      getTombstoneSweepStatus({ silent: true })
-        .then((r) => { if (!cancelled && Array.isArray(r?.refused)) setRefused(r.refused); })
-        .catch(() => { if (!cancelled) setRefused((prev) => prev ?? []); });
-    };
-    const refreshAfterPeerChange = () => {
-      fetchStatus();
-      if (backfillTimer) clearTimeout(backfillTimer);
-      backfillTimer = setTimeout(() => {
-        backfillTimer = null;
-        if (!cancelled) fetchStatus();
-      }, 1500);
-    };
-    fetchStatus();
-    socket.on('instances:peers:updated', refreshAfterPeerChange);
-    return () => {
-      cancelled = true;
-      if (backfillTimer) clearTimeout(backfillTimer);
-      socket.off('instances:peers:updated', refreshAfterPeerChange);
-    };
-  }, []);
-
-  const [runSweep, sweeping] = useAsyncAction(async () => {
-    const result = await sweepTombstonesNow({ graceMs: 0 }, { silent: true });
-    if (!result) return null;
-    const { universes = 0, series = 0, issues = 0, collections = 0, refused: refusedKinds = [] } = result;
-    const totalPruned = universes + series + issues + collections;
-    if (totalPruned > 0) {
-      const parts = [];
-      if (universes) parts.push(`${universes} universe${universes === 1 ? '' : 's'}`);
-      if (series) parts.push(`${series} series`);
-      if (issues) parts.push(`${issues} issue${issues === 1 ? '' : 's'}`);
-      if (collections) parts.push(`${collections} collection${collections === 1 ? '' : 's'}`);
-      toast.success(`Pruned ${parts.join(' / ')}`);
-    } else if (refusedKinds.length === 0) {
-      toast('No tombstones eligible for pruning');
-    }
-    if (refusedKinds.length > 0) {
-      const refusedLabel = [...new Set(refusedKinds.map((k) => TOMBSTONE_KIND_PLURAL[k] ?? k))].join(', ');
-      toast.warning(`Refused: ${refusedLabel} — a snapshot-mode peer has no per-record subscription for this kind. Resolve to enable pruning.`);
-    }
-    setRefused(refusedKinds);
-    return result;
-  }, { errorMessage: 'Tombstone sweep failed' });
-
-  // Every cohort refused iff `universe` (its own cohort), `series` (the
-  // pipeline cohort representative — issues always travel with series), AND
-  // `mediaCollection` (its own cohort) are all in the refused set.
-  const allRefused = refused != null
-    && refused.includes('universe')
-    && refused.includes('series')
-    && refused.includes('mediaCollection');
-  const disabled = sweeping || refused == null || allRefused;
-  const title = allRefused
-    ? 'Every kind has an enabled snapshot-mode peer with no per-record subscription — pruning would risk resurrection. Subscribe the peer or disable it to enable GC.'
-    : 'Prune tombstones acked by every subscribed peer (skips the 24h grace).';
-
-  return (
-    <div className="bg-port-card border border-port-border rounded-xl p-4 flex items-center justify-between gap-3">
-      <div className="flex items-center gap-2 text-sm text-gray-300">
-        <Trash2 size={14} className="text-gray-500" />
-        <span className="text-gray-400 uppercase tracking-wider text-xs font-medium">Tombstone GC</span>
-        <span className="text-gray-500 text-xs">
-          Run on demand instead of waiting for the 24h orchestrator sweep.
-        </span>
-      </div>
-      <button
-        onClick={runSweep}
-        disabled={disabled}
-        title={title}
-        className="bg-port-accent hover:bg-port-accent/80 disabled:opacity-40 disabled:cursor-not-allowed text-white px-3 py-1.5 rounded text-xs font-medium transition-colors flex items-center gap-1.5"
-      >
-        <Trash2 size={12} />
-        {sweeping ? 'Pruning…' : 'GC tombstones now'}
-      </button>
-    </div>
-  );
-}
-
 export default function Instances() {
   const [self, setSelf] = useState(null);
   const [peers, setPeers] = useState([]);
   const [syncStatus, setSyncStatus] = useState(null);
   const [tailnetInfo, setTailnetInfo] = useState(null);
+  const [networkExposure, setNetworkExposure] = useState(null);
   const [loading, setLoading] = useState(true);
 
   const fetchData = useCallback(async () => {
@@ -1083,17 +1015,18 @@ export default function Instances() {
       setPeers(data.peers);
       setSyncStatus(data.syncStatus ?? null);
     }
-    setLoading(false);
   }, []);
 
   useEffect(() => {
-    // Kick off both reads in parallel — the tailnet map is independent of
-    // peer state, and fetching it sequentially added ~100-500ms to first paint
-    // on machines where `tailscale status --json` is slow to respond.
+    // Kick off these reads in parallel — the tailnet map and exposure snapshot
+    // are independent of peer state, and fetching them sequentially added
+    // ~100-500ms to first paint on machines where `tailscale status --json`
+    // is slow to respond.
     Promise.all([
       fetchData(),
-      getTailnetInfo().then(setTailnetInfo).catch(() => setTailnetInfo(null))
-    ]);
+      getTailnetInfo().then(setTailnetInfo).catch(() => setTailnetInfo(null)),
+      getNetworkExposure({ silent: true }).then(setNetworkExposure).catch(() => setNetworkExposure(null))
+    ]).finally(() => setLoading(false));
 
     socket.emit('instances:subscribe');
     const handlePeersUpdated = (updatedPeers) => {
@@ -1128,14 +1061,12 @@ export default function Instances() {
         <span className="text-sm text-gray-500">PortOS Federation</span>
       </div>
 
-      <TailnetHelpBanner tailnetInfo={tailnetInfo} />
+      <TailnetHelpBanner tailnetInfo={tailnetInfo} networkExposure={networkExposure} />
 
       <div className="grid grid-cols-1 md:grid-cols-2 gap-4">
         <SelfCard self={self} onUpdate={fetchData} syncStatus={syncStatus} tailnetInfo={tailnetInfo} />
         <AddPeerForm onAdd={fetchData} />
       </div>
-
-      <TombstoneGcSection />
 
       {peers.length > 0 && (
         <div>
